Updated test suite
